pyproject.toml
[project]
name = "picloud"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "hsn", email = "hsn1919810@gmail.com" },
]
dependencies = [
    "fastapi>=0.109.2",
    "uvicorn[standard]>=0.27.0.post1",
    "python-multipart>=0.0.7",
    "requests>=2.31.0",
    "pillow>=10.2.0",
    "aiohttp>=3.9.3",
]
requires-python = "==3.11.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false

[tool.pdm.scripts]
#start = "uvicorn src.picloud.main:app --reload"
start = {call="src.picloud.main:app"}
